.github/tests/requirements-test.txt
pytest>=7.0
pytest-cov>=4.0
PyGithub>=1.59
GitPython>=3.1
requests>=2.28
